LPS-76947 portal-configuration-settings-test - extends tests from a common base test case so they can share updating logic
